renamed jcache sample infinispan-max-entries to infinispan-eviction sample for each memory configuration
