+++ jcache/infinispan-eviction/src/test/java/org/javaee8/jcache/infinispan/eviction/ObjectStorageTest.java
@@ -1,0 +1,25 @@
+package org.javaee8.jcache.infinispan.eviction;
+
+import org.jboss.arquillian.container.test.api.Deployment;
+import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.shrinkwrap.api.Archive;
+import org.jboss.shrinkwrap.api.ShrinkWrap;
+import org.jboss.shrinkwrap.api.spec.WebArchive;
+import org.junit.runner.RunWith;
+
+
+/**
+ * @author Radim Hanus
+ */
+@RunWith(Arquillian.class)
+public class ObjectStorageTest extends KeyValueServiceTest {
+    @Deployment
+    public static Archive<?> deploy() {
+        return ShrinkWrap.create(WebArchive.class)
+                .addClasses(KeyValueService.class, KeyValueServiceTest.class)
+                .addAsResource("infinispan-object.xml", "META-INF/infinispan.xml")
+                .addAsWebInfResource("beans.xml")
+                .addAsWebInfResource("jboss-deployment-structure.xml");
+    }
+
+}
