+++ jcache/infinispan-eviction/src/test/resources/beans.xml
