+++ jcache/infinispan-eviction/src/test/resources/jboss-deployment-structure.xml
